Add Fragment Attribute Interface.
This interface (FAI) makes it possible for plugins to provide
programmable per-fragment operations. Each plugin can add a set of
fragment functions to the fragment attributes that are used for
a drawing operation. A fragment function contains a set of data
defined by some existing extension to OpenGL that provide
programmable per-fragment operations. Current version of this
interface only support GL_ARB_fragment_program but it is designed
in such a way that it can very easily be extended to support other
OpenGL extensions.

The FAI also provide mechanisms which allow plugins to allocated
texture units and parameters for private use.

The fragment attribute implementation contains methods for
combining a set fragment functions into a fragment program
that can be used by OpenGL for per-fragment operations.

The major benefit of the FAI is that it provides a pluggable
way to use programmable per-fragment operations and will allow
us to take advantage of the latest OpenGL extensions and
create some really amazing effects while still making sure
that everything integrate perfectly.