Clean up and document window paint flags.
TRANSLUCENT flag that was earlier used for both indicating
that we're painting clipped translucent window regions and
that blending is required. The read-only CLIP_TRANSLUCENT
flag is now used to indicate that we're painting clipped
translucent window regions. TRANSLUCENT flag now indicate
that client window is translucent. BLEND flag indicates that
blending should be used when rendering.