gwd: Disable scaling support in GTK+3
We can't really support the GTK+3 automatic scaling, as to much
code relies on the GdkWindow and XWindow sizes, etc to match.
In order to keep working we just disable the scaling, meaning
we will pick up the larger fonts, but nothing else. Its not
ideal but it works for now.

https://bugzilla.gnome.org/show_bug.cgi?id=706388
https://git.gnome.org/browse/metacity/commit/?id=a1b96fd