resize: Do the final damage when we close up
Fixes #1

The XXX-outlines the basic problem, and it's a fundamental problem in the
resize plugin.

resizeTerminate doesn't actually disable drawing of the rectangle, so since
we just did damage there, we had a race condition. If we got a
paint-notification before we got the resize-one, we would still paint the
new rectangle.

This commit makes sure that the last thing we do after terminating the
resize is to damage. This means that even though we might actually paint
the rectangle between resizeTerminate and resizeFinishTerminate, it's of no
real concern.

The patch itself is correct and makes things better, but I'd like to clean
out this wacko logic all together eventually.