Move collision check to outside of loop
On top of this, this also allows global button bindings to be
distinguished from desktop-only button bindings. For example, the user
might want to annotate several windows with <Control><Alt>Button1.
However, they also want to bring up a window listing menu with
<ClickOnDesktop><Control><Alt>Button1. They'll be able to annotate
normally and be able to bring up a window menu if they want to with
<Control><Alt>Button1 depending on where they click.

That being said, the eventHitDesktop isn't picking this up in the case
of passthrough windows yet (eg: tint2). vpswitch is able to pick up on
this, but we'll need to check how vpswitch is able to pick that up.