Fix longstanding tooltip cutoff bug
Sometimes, tooltips would render with the size of the last tooltip,
causing a problem where tooltips were either cutoff or a shadow larger
than the tooltip texture size, noticeable with animations enabled.
The bug ended up having to do with keeping a pixmap that was unmapped
on the client side but not on the window manager side, specifically
for animations. The window size was not being updated and caused the
problem. Now we always update the size regardless if the pixmap is being
kept or not. Fixes #34.