Damage other outputs when displaying unredirected fullscreen window
When an unredirected fullscreen window is present on one output,
other outputs get no damage even if damageScreen() was called
because COMP_SCREEN_DAMAGE_ALL_MASK flag was unset when
overlayWindowCount > 0 on the CompScreen. Here we fix this by
only unsetting the flag when only a single output is present.