Use direct method when selecting output device for geometry
outputDeviceForGeometry() takes a rectangular region of the screen and returns
an output device number where the rectangle lives. This function was returning
incorrect output for rectangles outside of the screen area. This patch trims
the code that was previously causing the trouble, presumably in place for
overlapping outputs, which is typically uncommon. This bug is most notable
in grid with multiple outputs, where this function selects the output
and the window rect is sometimes offscreen.