Never maximize windows to another viewport
When maximizing a window, a viewport was chosen based on the center point
of the window to be maximized, and maximized to that viewport. This meant
that if your window is mostly on a viewport you can't see, maximizing it
effectively made the window disappear from the current viewport. This is
decidedly a bug and fixed by always choosing the current viewport for the
maximize target viewport.