Improve autoconf argument "default-plugins"
• Clear metadata/Makefile.am of a test that could be done earlier.
• Print a message during checking.
• Print a message with the result, if no error occurs.
• Add a help string, especially to indicate the expected syntax.

• Add a nested conditional structure to check the argument:

In the regular expression, * is used instead of + despite seeming
redundancy. This is done on purpose to restrict the nature of the
regular expression to a Basic Regular Expression (BRE), as POSIX
expr changed its matching expression syntax from Extended Regular
Expressions (in an early proposal) to BREs (in IEEE Std 1003.1)
"to avoid breaking historical applications".

http://pubs.opengroup.org/onlinepubs/9699919799/utilities/expr.html

Besides, using --with-default-plugins or --without-default-plugins
without any argument causes default_plugins to be assigned with the
value "yes" or "no", respectively. Although this is normal behaviour,
overwriting the value of default_plugins to "core" in such cases
avoids subsequent runtime errors.