Fix warpPointer() with XI2 events enabled
warpPointer() needs to filter out motion events from the queue to avoid
erroneous movements and deltas.  It does not currently filter XI2
events, and thus does not work properly when they are enabled.

For the moment, just disable XI2 motion events as we don't rely on them
yet and fixing warpPointer() to work with them is a lot more involved.