Update cursor when it gets changed from outside
When the cursor was changed from outside of compiz, it was kept till we
modify it on the compiz side (ie. through ccsm).
Now we listen to changes in the RESOURCE_MANAGER of the root Xorg window and then
apply the cursor-related changes we see (theme and size).
This is a stripped-down version of what is implemented in compiz 0.9.
The same is done on 3 places: core, gtk-window-decorator, move plugin.