Increase the maximum size of _NET_WM_ICON data we read
GTK might fill up to 262144 blocks, and this allows us to read larger
icons.  Doing so also has the nice side effect of supporting icon lists
where the first icon is larger than 65536 blocks (e.g. a 256*256 icon,
as it actually takes an extra 2 blocks for the icon dimensions, making
it 65538 blocks).

Size 262144 looks a bit arbitrary but it seems like a good value as
it's what GTK might fill up, and it allows way enough room for a list
containing all common sizes up to 256*256.

I am not entirely sure what happens if the internal X server maximum
size was 65536 which is the theoretical minimum, but all servers I can
test with support way more than this, and it doesn't seem worth it to
add the complexity needed for a more dynamic range.  If needed, it
could rely on X calls XMaxRequestSize() and XExtendedMaxRequestSize().